Remove readline and gdbm
readline and gdbm, which have GPLv3 license, are added as dependency
for many other software.

To allow build system without GPLv3, change build config for
these software to not use readline/gdbm.

Note that connman-client package is disabled with this change, but
it was listed in packagegroup-core-tools-testapps.
Has to removed there as well.

Signed-off-by: Hung Tran <hung.tran.jy@renesas.com>