Workaround to avoid GPLv3 hard dependency in rpm and libcheck
rpm-build package in rpm recipe RDEPENDS on debugedit, which has GPLv3 license.
To avoid GPLv3 dependency, forcefully remove RDEPENDS setting because
there is no clean config to disable this dependency.
As the result, there is chance that rpm-build is not-runnable,
or runnable but with issue.
(but there is no issue with other packages in rpm, *include rpm itself*).

libcheck RDEPENDS on gawk, which has GPLv3 license.
Replace this dependency by awk from busybox.
But note that busybox awk has some difference with gawk,
so issue can happen when libcheck try to use awk.

Signed-off-by: Hung Tran <hung.tran.jy@renesas.com>